Python Developer Course/requirements.txt
certifi==2022.6.15.1
charset-normalizer==2.1.1
coverage==6.4.4          
idna==3.3
Pillow==9.2.0
pony==0.7.16
psycopg2==2.9.3
pygame==2.1.2
requests==2.28.1
simple-draw==2.8.0
termcolor==2.0.1
urllib3==1.26.12
vk-api==11.9.9